+++ changelog.md
@@ -1,3 +1,7 @@
+# Version 0.2
+  - UL20xx Status byte ignored if 0xFF
+  - UL20xx decoder warnings more consistent.
+
 # Version 0.1
   - UL20xx decoder with different output formatting options
   - CM30xx decoder with different output formatting options